Moved the shorten study name function to utils and added unit tests
Moved the shorten study name function to utils and added unit tests
